fix: tests

+++ src/main/java/io/supertokens/webserver/api/multitenancy/RemoveTenantAPI.java
@@ -53,8 +53,6 @@ public class RemoveTenantAPI extends WebserverAPI {
         String tenantId = InputParser.parseStringOrThrowError(input, "tenantId", false);
         tenantId = Utils.normalizeAndValidateTenantId(tenantId);
 
-
-
         if (tenantId.equals(TenantIdentifier.DEFAULT_TENANT_ID)) {
             throw new ServletException(new BadPermissionException("Cannot delete public tenant, use remove app API instead"));
         }
@@ -62,7 +60,9 @@ public class RemoveTenantAPI extends WebserverAPI {
         try {
             TenantIdentifier sourceTenantIdentifier = this.getTenantIdentifier(req);
 
-            enforcePublicTenantAndGetPublicTenantStorage(req); // Enforce public tenant
+            if (!sourceTenantIdentifier.getTenantId().equals(TenantIdentifier.DEFAULT_TENANT_ID)) {
+                throw new BadPermissionException("Only the public tenantId is allowed to delete a tenant");
+            }
 
             boolean didExist = Multitenancy.deleteTenant(new TenantIdentifier(sourceTenantIdentifier.getConnectionUriDomain(), sourceTenantIdentifier.getAppId(), tenantId), main);
